Move status class into history tracker and fix comments
